fix: initialMargin

+++ src/markets/perps/interfaces/IOrderModule.sol
@@ -46,7 +46,7 @@ interface IOrderModule {
         uint128 accountId,
         uint128 marketId,
         int128 sizeDelta,
-        uint256 acceptablePrice
+        uint128 acceptablePrice
     )
         external;
